// src/app/match-data/match-data-types.ts
export type WinRatioData = {
  matchCount: number;
  firstPlayerWinCount: number;
  firstPlayerWinRate: number;
  secondPlayerWinCount: number;
  secondPlayerWinRate: number;
  drawCount: number;
  drawRate: number;
};

export type WinRatio = {
  match_count: number;
  first_player_win: number;
  second_player_win: number;
  draw: number;
};